Move unit tests outside subdir.
