+++ tests/test_shrink.py
@@ -2,7 +2,7 @@
 
 from edo.fitness import get_population_fitness
 from edo.operators import selection, shrink
-from edo.pdfs import Gamma, Normal, Poisson
+from edo.families import Gamma, Normal, Poisson
 from edo.population import create_initial_population
 
 from .util.parameters import SHRINK
